+++ stackslib/src/clarity_vm/tests/large_contract.rs
@@ -48,7 +48,7 @@ use crate::clarity_vm::database::MemoryBackingStore;
 use crate::util_lib::boot::boot_code_id;
 
 fn test_block_headers(n: u8) -> StacksBlockId {
-    StacksBlockId([n as u8; 32])
+    StacksBlockId([n; 32])
 }
 
 pub const TEST_BURN_STATE_DB_AST_PRECHECK: UnitTestBurnStateDB = UnitTestBurnStateDB {
@@ -131,7 +131,7 @@ fn test_simple_token_system(#[case] version: ClarityVersion, #[case] epoch: Stac
 
     let mut gb = clarity.begin_test_genesis_block(
         &StacksBlockId::sentinel(),
-        &StacksBlockId([0xfe as u8; 32]),
+        &StacksBlockId([0xfe; 32]),
         &TEST_HEADER_DB,
         burn_db,
     );
@@ -197,8 +197,8 @@ fn test_simple_token_system(#[case] version: ClarityVersion, #[case] epoch: Stac
     {
         let mut block = new_block(
             &mut clarity,
-            &StacksBlockId([0xfe as u8; 32]),
-            &StacksBlockId([0 as u8; 32]),
+            &StacksBlockId([0xfe; 32]),
+            &StacksBlockId([0; 32]),
             &TEST_HEADER_DB,
             burn_db,
         );
@@ -697,7 +697,7 @@ pub fn rollback_log_memory_test(
     clarity_instance
         .begin_test_genesis_block(
             &StacksBlockId::sentinel(),
-            &StacksBlockId([0 as u8; 32]),
+            &StacksBlockId([0; 32]),
             &TEST_HEADER_DB,
             burn_db,
         )
@@ -706,8 +706,8 @@ pub fn rollback_log_memory_test(
     {
         let mut conn = new_block(
             &mut clarity_instance,
-            &StacksBlockId([0 as u8; 32]),
-            &StacksBlockId([1 as u8; 32]),
+            &StacksBlockId([0; 32]),
+            &StacksBlockId([1; 32]),
             &TEST_HEADER_DB,
             burn_db,
         );
@@ -768,7 +768,7 @@ pub fn let_memory_test(#[case] clarity_version: ClarityVersion, #[case] epoch_id
     clarity_instance
         .begin_test_genesis_block(
             &StacksBlockId::sentinel(),
-            &StacksBlockId([0 as u8; 32]),
+            &StacksBlockId([0; 32]),
             &TEST_HEADER_DB,
             burn_db,
         )
@@ -777,8 +777,8 @@ pub fn let_memory_test(#[case] clarity_version: ClarityVersion, #[case] epoch_id
     {
         let mut conn = new_block(
             &mut clarity_instance,
-            &StacksBlockId([0 as u8; 32]),
-            &StacksBlockId([1 as u8; 32]),
+            &StacksBlockId([0; 32]),
+            &StacksBlockId([1; 32]),
             &TEST_HEADER_DB,
             burn_db,
         );
@@ -847,7 +847,7 @@ pub fn argument_memory_test(
     clarity_instance
         .begin_test_genesis_block(
             &StacksBlockId::sentinel(),
-            &StacksBlockId([0 as u8; 32]),
+            &StacksBlockId([0; 32]),
             &TEST_HEADER_DB,
             burn_db,
         )
@@ -856,8 +856,8 @@ pub fn argument_memory_test(
     {
         let mut conn = new_block(
             &mut clarity_instance,
-            &StacksBlockId([0 as u8; 32]),
-            &StacksBlockId([1 as u8; 32]),
+            &StacksBlockId([0; 32]),
+            &StacksBlockId([1; 32]),
             &TEST_HEADER_DB,
             burn_db,
         );
@@ -924,7 +924,7 @@ pub fn fcall_memory_test(#[case] clarity_version: ClarityVersion, #[case] epoch_
     clarity_instance
         .begin_test_genesis_block(
             &StacksBlockId::sentinel(),
-            &StacksBlockId([0 as u8; 32]),
+            &StacksBlockId([0; 32]),
             &TEST_HEADER_DB,
             burn_db,
         )
@@ -933,8 +933,8 @@ pub fn fcall_memory_test(#[case] clarity_version: ClarityVersion, #[case] epoch_
     {
         let mut conn = new_block(
             &mut clarity_instance,
-            &StacksBlockId([0 as u8; 32]),
-            &StacksBlockId([1 as u8; 32]),
+            &StacksBlockId([0; 32]),
+            &StacksBlockId([1; 32]),
             &TEST_HEADER_DB,
             burn_db,
         );
@@ -1043,7 +1043,7 @@ pub fn ccall_memory_test(#[case] clarity_version: ClarityVersion, #[case] epoch_
     clarity_instance
         .begin_test_genesis_block(
             &StacksBlockId::sentinel(),
-            &StacksBlockId([0 as u8; 32]),
+            &StacksBlockId([0; 32]),
             &TEST_HEADER_DB,
             burn_db,
         )
@@ -1052,8 +1052,8 @@ pub fn ccall_memory_test(#[case] clarity_version: ClarityVersion, #[case] epoch_
     {
         let mut conn = new_block(
             &mut clarity_instance,
-            &StacksBlockId([0 as u8; 32]),
-            &StacksBlockId([1 as u8; 32]),
+            &StacksBlockId([0; 32]),
+            &StacksBlockId([1; 32]),
             &TEST_HEADER_DB,
             burn_db,
         );
